+++ src/api/@types/responses.ts
@@ -1,0 +1,3 @@
+export interface Res500 {
+  error: string;
+}
